Refactor MassTransitHelpers and tests to extensions project
