+++ src/Agent/NewRelic/Agent/Extensions/Providers/Wrapper/MassTransitLegacy/NewRelicFilter.cs
@@ -6,6 +6,8 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using GreenPipes;
 using MassTransit;
+using NewRelic.Agent.Api;
+using NewRelic.Agent.Extensions.Helpers;
 using NewRelic.Agent.Extensions.Providers.Wrapper;
 using MethodCall = NewRelic.Agent.Extensions.Providers.Wrapper.MethodCall;
 
@@ -88,7 +90,7 @@ namespace NewRelic.Providers.Wrapper.MassTransitLegacy
             var queueData = MassTransitHelpers.GetQueueData(context.SourceAddress);
 
             var transaction = _agent.CurrentTransaction;
-            MassTransitHelpers.InsertDistributedTraceHeaders(context.Headers, transaction);
+            InsertDistributedTraceHeaders(context.Headers, transaction);
             var segment = transaction.StartMessageBrokerSegment(mc, queueData.DestinationType, MessageBrokerAction.Produce, MessageBrokerVendorName, queueData.QueueName);
 
             await next.Send(context);
@@ -105,11 +107,22 @@ namespace NewRelic.Providers.Wrapper.MassTransitLegacy
             var queueData = MassTransitHelpers.GetQueueData(context.SourceAddress);
 
             var transaction = _agent.CurrentTransaction;
-            MassTransitHelpers.InsertDistributedTraceHeaders(context.Headers, transaction);
+            InsertDistributedTraceHeaders(context.Headers, transaction);
             var segment = transaction.StartMessageBrokerSegment(mc, queueData.DestinationType, MessageBrokerAction.Produce, MessageBrokerVendorName, queueData.QueueName);
 
             await next.Send(context);
             segment.End();
         }
+
+        public static void InsertDistributedTraceHeaders(SendHeaders headers, ITransaction transaction)
+        {
+            var setHeaders = new Action<SendHeaders, string, string>((carrier, key, value) =>
+            {
+                carrier.Set(key, value);
+            });
+
+            transaction.InsertDistributedTraceHeaders(headers, setHeaders);
+        }
+
     }
 }
