+++ tests/Agent/UnitTests/NewRelic.Agent.Extensions.Tests/TestClasses.cs
@@ -1,7 +1,7 @@
 // Copyright 2020 New Relic, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-namespace Agent.Extensions.Test
+namespace Agent.Extensions.Tests
 {
     public class SimpleClass
     {
